Add usePackageTranslation
